ai/pyproject.toml
[project]
name = "lumineer-ai"
version = "0.1.0"
description = "Lumineer AI Processing Service"
requires-python = ">=3.12"
dependencies = [
    "litestar[standard]>=2.14",
    "openai>=1.60",
    "openai-agents>=0.1",
    "qdrant-client[fastembed]>=1.13",
    "pandas>=2.2",
    "pyarrow>=18.0",
    "pydantic-settings>=2.7",
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "en-core-web-sm",
    "tenacity>=9.0",
    "tqdm>=4.67",
    "uvicorn[standard]>=0.34",
    "langfuse>=2.0,<4.0",
    "prometheus-client>=0.21",
    # MCP Server (stretch goal)
    "mcp[cli]>=1.6",
    "PyJWT[cryptography]>=2.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scripts"]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }

[dependency-groups]
dev = [
    "ruff>=0.9",
    "mypy>=1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.27",
    "deepeval>=2.5",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disable_error_code = ["operator", "union-attr", "arg-type", "unused-ignore", "misc", "no-untyped-call", "no-untyped-def"]

[[tool.mypy.overrides]]
module = "evals.*"
disallow_untyped_defs = false
disable_error_code = ["attr-defined", "no-untyped-call", "no-untyped-def"]

[tool.pytest.ini_options]
testpaths = ["tests"]
